Refactor submenu behavior based on feedback
